requirements.txt
requests
trading-ig
pandas
numpy
pytz
lightstreamer-client
munch
python-dotenv